moved DataOutAtomData and ParseAtomData to /atom
