+++ tests/atom/write_atom_data_into_vtp_01.cc
@@ -6,8 +6,8 @@
 #include <deal.II/grid/grid_out.h>
 
 #include <dealiiqc/atom/atom_handler.h>
-#include <dealiiqc/io/data_out_atom_data.h>
-#include <dealiiqc/io/configure_qc.h>
+#include <dealiiqc/atom/data_out_atom_data.h>
+#include <dealiiqc/configure/configure_qc.h>
 #include <dealiiqc/utilities.h>
 
 using namespace dealii;
